// planning-idsdp-rms-frontend/src/app/main/modules/rpm/models/ViewPresentationEvaluatorsFeedback.ts
export class ViewPresentationEvaluatorsFeedback {

    id: number;
    uuid: string;
    m2CreateSeminarId: number;
    m1ResearcherProposalId: number;
    concernedPersonUserId: number;
    stProfileOfExpertEvaluatorsId: number;
    m2ResearcherPresentationId: number;
    m2EvaluatorsFeedbackId: number;
    isPresent: number;
    expertEvaluatorDto: any;

}